+++ src/test/scala/tabulae/fstforms/FstIrregularFormSpec.scala
@@ -6,9 +6,30 @@ import org.scalatest.FlatSpec
 class FstIrregularFormSpec extends FlatSpec {
 
 
-  "The Form object" should "construct a parsed noun form from FST string input for irregular nouns" in pending
+  "The Form object" should "construct a parsed noun form from FST string input for irregular nouns" in {
+    val fst = "<u>ocremorph.n25359mns</u><u>lexent.n25359</u>ivppiter<masc><nom><sg><irregnoun><div><irregnoun><u>irreginfl.0</u>"
+    val f = LemmatizedForm(fst)
+    f match {
+      case nf: NounForm => {
+        assert (nf.gender == Masculine)
+    assert (nf.grammaticalCase == Nominative)
+        assert (nf.grammaticalNumber == Singular)
+      }
+      case _ => fail("Should have created a noun form")
+    }
+  }
   it should "construct a parsed adjective form from FST string input for irregular adjectives" in pending
-  it should "construct a parsed adverb form from FST string input for irregular adverbs" in pending // <u>lat25morph.advn790</u><u>ls.n790</u>adhuc<pos><irregadv><div><irregadv><u>irreginfl.2</u>
+  it should "construct a parsed adverb form from FST string input for irregular adverbs" in {
+    val fst = "<u>ocremorph.n25115</u><u>ls.n25115</u>itervm<pos><irregadv><div><irregadv><u>irreginfl.2</u>"
+    val f = LemmatizedForm(fst)
+    f match {
+      case advf: AdverbForm => {
+        assert (advf.degree == Positive)
+      }
+      case _ => fail("Should have created an adverb form")
+    }
+  }
+  
   it should "construct a parsed infinitive form from FST string input for irregular infinitive" in pending
   it should "construct a parsed participle form from FST string input for irregular participle" in pending
   it should "construct a parsed verb form from FST string input for irregular verb" in pending